playwright-python/requirements.txt
pytest
playwright0